requirements.txt
gpt4free
fastapi
uvicorn[standard]
python-dotenv
pydantic